requirements.txt
aiofiles~=22.1.0
apscheduler>=3.8,<4 # from apscheduler~=3.8.0
hikari[speedups]==2.0.0.dev111 
hikari-lightbulb>=2.2.1,<3 
isodate~=0.6.1 # for date parsing
psutil>=5.8,<6 # for the stats plugin
pygount>=1.2,<2 # for the stats command
python-dotenv~=0.21.0 # for dotenv support
rapidfuzz>=2.0.7,<3 # for fuzzy string matching
uvloop~=0.17.0; os_name != "nt" # uvloop is not supported on Windows
aiohttp~=3.8.3 # for the web server
pytz~=2022.4 # for timezone support
miru~=0.2.0
setuptools~=58.0.4
numpy~=1.23.4